requirements.txt
requests>=2.31
curl_cffi>=0.7
playwright>=1.40
discord.py>=2.3
python-dotenv>=1.0
anthropic>=0.39
fastapi>=0.115
uvicorn>=0.32
